refactor(WeaponStats): Range is now stored as a float for a wider variety

+++ Endless_Survivor/Assets/Scripts/Weapons/WeaponStats.cs
@@ -10,13 +10,13 @@ public class WeaponStats
     static readonly float _atkSpeedStatVariation = .15f;
     static readonly float _damageStatVariation = 1f;
 
-    [SerializeField] int _range;
+    [SerializeField] float _range;
     [SerializeField] float _attackSpeed;
     [SerializeField] int _damage;
     int _trueLevel;
     int _inducedLevel = 0;
 
-    public int Range { get { return _range; } }
+    public float Range { get { return _range; } }
     public float AttackSpeed { get { return _attackSpeed; } }
     public int Damage { get { return _damage; } }
     public int TrueLevel { get { return _trueLevel; } }
@@ -35,14 +35,14 @@ public class WeaponStats
     public void SetTrueLevelStats(WeaponStats statsScaling, int level)
     {
         _trueLevel = level;
-        _range += (int) GetTrueStatIncrease(statsScaling.Range, level, _range, _rangeStatVariation);
+        _range += GetTrueStatIncrease(statsScaling.Range, level, _range, _rangeStatVariation);
         _attackSpeed += GetTrueStatIncrease(statsScaling.AttackSpeed, level, _attackSpeed, _atkSpeedStatVariation);
         _damage += (int) GetTrueStatIncrease(statsScaling.Damage, level, _damage, _damageStatVariation);
     }
     public void InducedLevelUp(WeaponStats statsScaling)
     {
         _inducedLevel++;
-        _range += (int)GetInducedStatIncrease(statsScaling.Range, _trueLevel + _inducedLevel, _range, _rangeStatVariation);
+        _range += GetInducedStatIncrease(statsScaling.Range, _trueLevel + _inducedLevel, _range, _rangeStatVariation);
         _attackSpeed += (int)GetInducedStatIncrease(statsScaling.AttackSpeed, _trueLevel + _inducedLevel, _attackSpeed, _atkSpeedStatVariation);
         _damage += (int)GetInducedStatIncrease(statsScaling.Damage, _trueLevel + _inducedLevel, _damage, _damageStatVariation);
     }
